requirements.txt
absl-py==2.0.0
addict==2.4.0
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==4.2.4
captcha==0.4
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
humanfriendly==10.0
idna==3.4
IMcore==0.2.7
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.15.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.0.9
pyreadline3==3.4.1
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
QtFusion==0.5.4
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.16.2
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
shiboken6==6.5.1.1
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.3
validators==0.22.0
watchdog==3.0.0
Werkzeug==3.0.1
zipp==3.17.0
